requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.1
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.1.4
distro==1.9.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
md2tgmd==0.3.10
mistletoe==1.4.0
multidict==6.7.0
openai==2.15.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.45
telegramify-markdown==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
watchdog==6.0.0
yarl==1.22.0
